+++ src/main/java/com/solvd/lab/army/dao/impl/BaseDAOImpl.java
@@ -40,7 +40,7 @@ public class BaseDAOImpl implements IBaseDAO  {
 
 
     @Override
-    public Base getById(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public Base getById(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Base base = null;
         Connection connection = null;
         ResultSet resultSet = null;
@@ -48,7 +48,7 @@ public class BaseDAOImpl implements IBaseDAO  {
         try {
             connection = connectionPool.getInstance().getConnection();
             statement = connection.prepareStatement(SELECT_BY_ID_QUERY);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             resultSet = statement.executeQuery();
             if (resultSet.next()) {
                 base = getDataFromResultSet(resultSet);
@@ -140,7 +140,7 @@ public class BaseDAOImpl implements IBaseDAO  {
     }
 
     @Override
-    public void delete(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public void delete(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Connection connection = null;
         PreparedStatement statement = null;
         try {
