+++ src/main/java/com/solvd/lab/army/dao/impl/CityDAOImpl.java
@@ -38,7 +38,7 @@ public class CityDAOImpl implements ICityDAO {
 
 
     @Override
-    public City getById(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public City getById(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         City city = null;
         Connection connection = null;
         ResultSet resultSet = null;
@@ -47,7 +47,7 @@ public class CityDAOImpl implements ICityDAO {
 
             connection = connectionPool.getInstance().getConnection();
             statement = connection.prepareStatement(SELECT_BY_ID_QUERY);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             resultSet = statement.executeQuery();
             if (resultSet.next()) {
                 city = getDataFromResultSet(resultSet);
@@ -134,7 +134,7 @@ public class CityDAOImpl implements ICityDAO {
     }
 
     @Override
-    public void delete(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public void delete(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Connection connection = null;
         PreparedStatement statement = null;
         try {
