+++ src/main/java/com/solvd/lab/army/dao/impl/EquipmentTypeDAOImpl.java
@@ -38,7 +38,7 @@ public class EquipmentTypeDAOImpl implements IEquipmentTypeDAO {
 
 
     @Override
-    public EquipmentType getById(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public EquipmentType getById(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         EquipmentType equipmentType = null;
         Connection connection = null;
         ResultSet resultSet = null;
@@ -46,7 +46,7 @@ public class EquipmentTypeDAOImpl implements IEquipmentTypeDAO {
         try {
             connection = connectionPool.getInstance().getConnection();
             statement = connection.prepareStatement(SELECT_BY_ID_QUERY);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             resultSet = statement.executeQuery();
             if (resultSet.next()) {
                 equipmentType = getDataFromResultSet(resultSet);
@@ -129,7 +129,7 @@ public class EquipmentTypeDAOImpl implements IEquipmentTypeDAO {
     }
 
     @Override
-    public void delete(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public void delete(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Connection connection = null;
         PreparedStatement statement = null;
         try {
