+++ src/main/java/com/solvd/lab/army/dao/impl/LocationDAOImpl.java
@@ -37,7 +37,7 @@ public class LocationDAOImpl implements ILocationDAO {
 
 
     @Override
-    public Location getById(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public Location getById(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Location location = null;
         Connection connection = null;
         ResultSet resultSet = null;
@@ -45,7 +45,7 @@ public class LocationDAOImpl implements ILocationDAO {
         try {
             connection = connectionPool.getInstance().getConnection();
             statement = connection.prepareStatement(SELECT_BY_ID_QUERY);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             resultSet = statement.executeQuery();
             if (resultSet.next()) {
                 location = getDataFromResultSet(resultSet);
@@ -130,7 +130,7 @@ public class LocationDAOImpl implements ILocationDAO {
     }
 
     @Override
-    public void delete(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public void delete(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Connection connection = null;
         PreparedStatement statement = null;
         try {
