+++ src/main/java/com/solvd/lab/army/dao/impl/RankTypeDAOImpl.java
@@ -36,7 +36,7 @@ public class RankTypeDAOImpl implements IRankTypeDAO {
 
 
     @Override
-    public RankType getById(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public RankType getById(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         RankType rankType = null;
         Connection connection = null;
         ResultSet resultSet = null;
@@ -44,7 +44,7 @@ public class RankTypeDAOImpl implements IRankTypeDAO {
         try {
             connection = connectionPool.getInstance().getConnection();
             statement = connection.prepareStatement(SELECT_BY_ID_QUERY);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             resultSet = statement.executeQuery();
             if (resultSet.next()) {
                 rankType = getDataFromResultSet(resultSet);
@@ -127,7 +127,7 @@ public class RankTypeDAOImpl implements IRankTypeDAO {
     }
 
     @Override
-    public void delete(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public void delete(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Connection connection = null;
         PreparedStatement statement = null;
         try {
