+++ src/main/java/com/solvd/lab/army/dao/impl/RoleDAOImpl.java
@@ -33,7 +33,7 @@ public class RoleDAOImpl implements IRoleDAO {
 
 
     @Override
-    public Role getById(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public Role getById(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Role role = null;
         Connection connection = null;
         ResultSet resultSet = null;
@@ -41,7 +41,7 @@ public class RoleDAOImpl implements IRoleDAO {
         try {
             connection = connectionPool.getInstance().getConnection();
             statement = connection.prepareStatement(SELECT_BY_ID_QUERY);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             resultSet = statement.executeQuery();
             if (resultSet.next()) {
                 role = getDataFromResultSet(resultSet);
@@ -124,7 +124,7 @@ public class RoleDAOImpl implements IRoleDAO {
     }
 
     @Override
-    public void delete(int id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
+    public void delete(long id) throws SQLException, IOException, InterruptedException, ClassNotFoundException {
         Connection connection = null;
         PreparedStatement statement = null;
         try {
